ViewRenderer instead of ContentViewHandler
