--- RiveMaui/RiveView.cs
@@ -1,12 +1,0 @@
-namespace RiveMaui;
-
-public class RiveView : ContentView
-{
-    public void Load(string animation)
-    {
-        if (Handler is RiveViewHandler handler)
-        {
-            handler.Load(animation);
-        }
-    }
-}
