+++ RiveMaui/RiveView/RiveView.cs
@@ -1,0 +1,12 @@
+namespace RiveMaui;
+
+public class RiveView : View
+{
+    public void Load(string animation)
+    {
+        if (Handler is RiveViewRenderer renderer)
+        {
+            renderer.Load(animation);
+        }
+    }
+}
