+++ RiveMaui/RiveView/RiveViewRenderer.ios.cs
@@ -1,0 +1,33 @@
+#if IOS
+using Foundation;
+using Microsoft.Maui.Controls.Handlers.Compatibility;
+using Rive.iOS;
+using UIKit;
+
+namespace RiveMaui;
+
+public partial class RiveViewRenderer : ViewRenderer<RiveView, UIView>
+{
+    public void Load(string animation)
+    {
+        var riveVM = new RiveViewModel(
+            animation,
+            ".riv",
+            NSBundle.MainBundle,
+            null,
+            RiveFit.contain,
+            RiveAlignment.center,
+            true,
+            null,
+            true,
+            null
+        );
+
+        var riveView = riveVM.CreateRiveView;
+        riveView.Frame = Element!.Bounds;
+
+        SetNativeControl(riveView);
+    }
+}
+
+#endif
